refactor(bird-adapter): use FQN versioned proto package
Migrate the bird-adapter operator to a fully-qualified, versioned proto
package (operators.bird_adapter.adapterpb.v1) so shared services such as
MetricsService and a future readiness service no longer collide on the
gateway. The directory hyphen prevents a matching package segment, so
PACKAGE_DIRECTORY_MATCH is disabled for this path only.
