+++ operators/bird-adapter/adapterpb/meson.build
@@ -1,4 +1,4 @@
-proto_dir = meson.current_source_dir()
+proto_dir = join_paths(meson.current_source_dir(), 'v1')
 root_dir = meson.project_source_root()
 
 proto_files = [join_paths(proto_dir, 'bird_import.proto')]
