+++ operators/bird-adapter/adapterpb/v1/bird_import.proto
@@ -1,10 +1,10 @@
 syntax = "proto3";
 
-package adapterpb;
+package operators.bird_adapter.adapterpb.v1;
 
 import "common/commonpb/ipaddr.proto";
 
-option go_package = "github.com/yanet-platform/yanet2/operators/bird-adapter/adapterpb;adapterpb";
+option go_package = "github.com/yanet-platform/yanet2/operators/bird-adapter/adapterpb/v1;adapterpb";
 
 // AdapterService provides configuration for the BIRD import adapter.
 service AdapterService {
