+++ operators/bird-adapter/service.go
@@ -18,7 +18,7 @@ import (
 	"google.golang.org/grpc/encoding/gzip"
 
 	"github.com/yanet-platform/yanet2/modules/route-mpls/controlplane/routemplspb/v1"
-	adapterpb "github.com/yanet-platform/yanet2/operators/bird-adapter/adapterpb"
+	adapterpb "github.com/yanet-platform/yanet2/operators/bird-adapter/adapterpb/v1"
 	"github.com/yanet-platform/yanet2/operators/bird-adapter/internal/bird"
 	"github.com/yanet-platform/yanet2/operators/bird-adapter/internal/mpls"
 	"github.com/yanet-platform/yanet2/operators/bird-adapter/internal/rib"
